Фикс фантомных символов
